Project restructure & build.rs improvements
- collapse package into single crate `sqltk`

- build.rs

  - detect presence of cargo-expand and if install if missing.

  - run MetadataCommand in tempdir to retain pristine Cargo.lock and
    avoid warning during `cargo package` and `cargo publish` commands.

  - auto-detect sqlparser package installation so that codegen will
    resolve against the resolved version when sqltk is used in another
    project.
