+++ src/visitable_blanket_impls.rs
@@ -1,4 +1,5 @@
 use sqlparser::ast::{OneOrManyWithParens, WrappedCollection};
+use visitor_helper::visit;
 
 use crate::*;
 
@@ -76,22 +77,4 @@ where
             WrappedCollection::Parentheses(nodes) => nodes.accept(visitor),
         }
     }
-}
-
-/// Helper function used by generated code for visiting a node and its children
-/// recursively while properly handling visitor control flow.
-#[inline(always)]
-pub(crate) fn visit<'ast, N, F, V>(
-    node: &'ast N,
-    visitor: &mut V,
-    visit_children: F,
-) -> ControlFlow<Break<V::Error>>
-where
-    V: Visitor<'ast>,
-    N: Visitable,
-    F: Fn(&mut V) -> ControlFlow<Break<V::Error>>,
-{
-    visitor.enter(node)?;
-    visit_children(visitor)?;
-    visitor.exit(node)
 }
